+++ CHANGES.md
@@ -1,2 +1,7 @@
+- 
+- Added debug mode to destroy cookie (with php)
+- Only show privacy policy url setting if cookie consent text is disabled
+- Added settings for cookie consent text and button
+
 0.1
 - Init